+++ CHANGES.md
@@ -1,6 +1,7 @@
 # UNRELEASED
 - [NEW] Allow a custom HTTP agent to be specified.
 - [NEW] Trim the database URL if it includes the `/_changes` endpoint.
+- [FIXED] Correctly pass request parameters to `feed.filter` function.
 - [FIXED] Retry `/_db_updates` request using the last recorded sequence value as
   the `since` query parameter.
 - [NOTE] Update engines in preparation for Node.js 4 “Argon” end-of-life.